scripts/requirements.txt
cryptography
piexif
Pillow

